fix(go): set the go directive to 1.22 (#87)
The go directive from the `go.mod` file should be set to the minimum
version required to compile packages from the module. This project's
code should work fine with any version greater than 1.22. This avoids
projects that depend on this package getting a go directive version bump
every time there's a new version of this dependency.